# Faster to import std.regex, by lazy eval of wordMatcher