After:
▶ time dmd test_case.d
dmd test_case.d  0.24s user 0.03s system 98% cpu 0.269 total

▶ gco master
Switched to branch 'master'

Before:
▶ time dmd test_case.d
dmd test_case.d  1.10s user 0.12s system 99% cpu 1.228 total