@CyberShadow 2017-12-22
Before:
![](https://dump.thecybershadow.net/3a13f9e8e0ee397cf8c9cdc53d150ad9/out.png)
After:
![](https://dump.thecybershadow.net/dca8eed8a252488f15bd70a665df8ec5/out.png)

Awesome!