@quickfur 2017-12-22
Looks like `std.uni` is the next candidate for improvement. :-P