@andralex 2017-12-22
@CyberShadow would be so nice to have that kind of information nicely visible on the website! "The upfront time cost of importing std modules"